requirements.txt
docutils==0.14
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.2.1